adding DimensionType to type_enums w/ linting fixes
