+++ dbt_semantic_interfaces/validations/validator_helpers.py
@@ -13,7 +13,6 @@ from pydantic import BaseModel, Extra
 
 from dbt_semantic_interfaces.enum_extension import assert_values_exhausted
 from dbt_semantic_interfaces.objects.base import FrozenBaseModel
-from dbt_semantic_interfaces.objects.elements.dimension import DimensionType
 from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.objects.semantic_manifest import SemanticManifest
 from dbt_semantic_interfaces.references import (
@@ -21,6 +20,7 @@ from dbt_semantic_interfaces.references import (
     SemanticModelElementReference,
     SemanticModelReference,
 )
+from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 
 VALIDATE_SAFELY_ERROR_STR_TMPLT = ". Issue occurred in method `{method_name}` called with {arguments_str}"
 ValidationContextJSON = Dict[str, Union[str, int, None]]
